Increase some test timeouts. Move tests around so the long pull test file is now ~100s
