+++ tests/SpeechRecognizerTests.ts
@@ -1254,6 +1254,7 @@ describe.each([true])("Service based tests", (forceNodeWebSocket: boolean) => {
                 done.fail(error);
             });
     }, 120000);
+
     test("PullStreamHalfFill", (done: jest.DoneCallback) => {
         // tslint:disable-next-line:no-console
         console.info("Name: PullStreamHalfFill");
@@ -1318,179 +1319,6 @@ describe.each([true])("Service based tests", (forceNodeWebSocket: boolean) => {
             (error: string) => {
                 done.fail(error);
             });
-    });
-
-    describe("Intiial Silence Tests", () => {
-        test("InitialSilenceTimeout (pull)", (done: jest.DoneCallback) => {
-            // tslint:disable-next-line:no-console
-            console.info("Name: InitialSilenceTimeout (pull)");
-            let p: sdk.PullAudioInputStream;
-            let bytesSent: number = 0;
-
-            // To make sure we don't send a ton of extra data.
-            // For reference, before the throttling was implemented, we sent 6-10x the required data.
-            const startTime: number = Date.now();
-
-            p = sdk.AudioInputStream.createPullStream(
-                {
-                    close: () => { return; },
-                    read: (buffer: ArrayBuffer): number => {
-                        bytesSent += buffer.byteLength;
-                        return buffer.byteLength;
-                    },
-                });
-
-            const config: sdk.AudioConfig = sdk.AudioConfig.fromStreamInput(p);
-
-            testInitialSilenceTimeout(config, done, (): void => {
-                const elapsed: number = Date.now() - startTime;
-
-                // We should have sent 5 seconds of audio unthrottled and then 2x the time reco took until we got a response.
-                const expectedBytesSent: number = (5 * 16000 * 2) + (2 * elapsed * 32000 / 1000);
-                expect(bytesSent).toBeLessThanOrEqual(expectedBytesSent);
-
-            });
-        }, 15000);
-
-        test("InitialSilenceTimeout (push)", (done: jest.DoneCallback) => {
-            // tslint:disable-next-line:no-console
-            console.info("Name: InitialSilenceTimeout (push)");
-            const p: sdk.PushAudioInputStream = sdk.AudioInputStream.createPushStream();
-            const bigFileBuffer: Uint8Array = new Uint8Array(1024 * 1024);
-            const config: sdk.AudioConfig = sdk.AudioConfig.fromStreamInput(p);
-
-            p.write(bigFileBuffer.buffer);
-            p.close();
-
-            testInitialSilenceTimeout(config, done);
-        }, 15000);
-
-        Settings.testIfDOMCondition("InitialSilenceTimeout (File)", (done: jest.DoneCallback) => {
-            // tslint:disable-next-line:no-console
-            console.info("Name: InitialSilenceTimeout (File)");
-            const audioFormat: AudioStreamFormatImpl = sdk.AudioStreamFormat.getDefaultInputFormat() as AudioStreamFormatImpl;
-            const bigFileBuffer: Uint8Array = new Uint8Array(1024 * 1024);
-            const bigFile: File = ByteBufferAudioFile.Load([audioFormat.header, bigFileBuffer.buffer]);
-
-            const config: sdk.AudioConfig = sdk.AudioConfig.fromWavFileInput(bigFile);
-
-            testInitialSilenceTimeout(config, done);
-        }, 15000);
-
-        const testInitialSilenceTimeout = (config: sdk.AudioConfig, done: jest.DoneCallback, addedChecks?: () => void): void => {
-            const s: sdk.SpeechConfig = BuildSpeechConfig();
-            objsToClose.push(s);
-
-            const r: sdk.SpeechRecognizer = new sdk.SpeechRecognizer(s, config);
-            objsToClose.push(r);
-
-            expect(r).not.toBeUndefined();
-            expect(r instanceof sdk.Recognizer);
-
-            let numReports: number = 0;
-
-            r.canceled = (o: sdk.Recognizer, e: sdk.SpeechRecognitionCanceledEventArgs) => {
-                done.fail(e.errorDetails);
-            };
-
-            r.recognized = (o: sdk.Recognizer, e: sdk.SpeechRecognitionEventArgs) => {
-                try {
-                    const res: sdk.SpeechRecognitionResult = e.result;
-                    expect(res).not.toBeUndefined();
-                    expect(sdk.ResultReason.NoMatch).toEqual(res.reason);
-                    expect(res.text).toBeUndefined();
-                    expect(res.properties).not.toBeUndefined();
-                    expect(res.properties.getProperty(sdk.PropertyId.SpeechServiceResponse_JsonResult)).not.toBeUndefined();
-
-                    const nmd: sdk.NoMatchDetails = sdk.NoMatchDetails.fromResult(res);
-                    expect(nmd.reason).toEqual(sdk.NoMatchReason.InitialSilenceTimeout);
-                } catch (error) {
-                    done.fail(error);
-                } finally {
-                    numReports++;
-                }
-
-            };
-
-            r.recognizeOnceAsync(
-                (p2: sdk.SpeechRecognitionResult) => {
-                    try {
-                        const res: sdk.SpeechRecognitionResult = p2;
-                        numReports++;
-
-                        expect(res).not.toBeUndefined();
-                        expect(sdk.ResultReason.NoMatch).toEqual(res.reason);
-                        expect(res.errorDetails).toBeUndefined();
-                        expect(res.text).toBeUndefined();
-                        expect(res.properties).not.toBeUndefined();
-                        expect(res.properties.getProperty(sdk.PropertyId.SpeechServiceResponse_JsonResult)).not.toBeUndefined();
-
-                        const nmd: sdk.NoMatchDetails = sdk.NoMatchDetails.fromResult(res);
-                        expect(nmd.reason).toEqual(sdk.NoMatchReason.InitialSilenceTimeout);
-                    } catch (error) {
-                        done.fail(error);
-                    }
-                },
-                (error: string) => {
-                    fail(error);
-                });
-
-            WaitForCondition(() => (numReports === 2), () => {
-                try {
-                    if (!!addedChecks) {
-                        addedChecks();
-                    }
-                    done();
-                } catch (error) {
-                    done.fail(error);
-                }
-            });
-        };
-
-        test.skip("InitialBabbleTimeout", (done: jest.DoneCallback) => {
-            // tslint:disable-next-line:no-console
-            console.info("Name: InitialBabbleTimeout");
-
-            const s: sdk.SpeechConfig = sdk.SpeechConfig.fromSubscription(Settings.SpeechSubscriptionKey, Settings.SpeechRegion);
-            expect(s).not.toBeUndefined();
-
-            s.speechRecognitionLanguage = "es-MX";
-
-            const config: sdk.AudioConfig = WaveFileAudioInput.getAudioConfigFromFile(Settings.WaveFile);
-
-            const r: sdk.SpeechRecognizer = new sdk.SpeechRecognizer(s, config);
-            expect(r).not.toBeUndefined();
-            expect(r instanceof sdk.Recognizer);
-
-            r.canceled = (o: sdk.Recognizer, e: sdk.SpeechRecognitionCanceledEventArgs): void => {
-                try {
-                    expect(e.errorDetails).toBeUndefined();
-                } catch (error) {
-                    done.fail(error);
-                }
-            };
-
-            r.recognizeOnceAsync(
-                (p2: sdk.SpeechRecognitionResult) => {
-                    try {
-                        const res: sdk.SpeechRecognitionResult = p2;
-                        expect(res).not.toBeUndefined();
-                        expect("What's the weather like?").toEqual(res.text);
-                        expect(res.reason).toEqual(sdk.ResultReason.RecognizedSpeech);
-
-                        r.close();
-                        s.close();
-                        done();
-                    } catch (error) {
-                        done.fail(error);
-                    }
-                },
-                (error: string) => {
-                    r.close();
-                    s.close();
-                    done.fail(error);
-                });
-        });
     });
 
     test.skip("emptyFile", (done: jest.DoneCallback) => {
@@ -1615,56 +1443,6 @@ describe.each([true])("Service based tests", (forceNodeWebSocket: boolean) => {
             });
     });
 
-    test("burst of silence", (done: jest.DoneCallback) => {
-        // tslint:disable-next-line:no-console
-        console.info("Name: burst of silence");
-        const s: sdk.SpeechConfig = BuildSpeechConfig();
-        objsToClose.push(s);
-
-        const f: ArrayBuffer = WaveFileAudioInput.LoadArrayFromFile(Settings.WaveFile);
-        const p: sdk.PushAudioInputStream = sdk.AudioInputStream.createPushStream();
-        const config: sdk.AudioConfig = sdk.AudioConfig.fromStreamInput(p);
-
-        const emptyBuffer: Uint8Array = new Uint8Array(1 * 1024);
-        p.write(emptyBuffer.buffer);
-        p.close();
-
-        const r: sdk.SpeechRecognizer = new sdk.SpeechRecognizer(s, config);
-        objsToClose.push(r);
-        expect(r).not.toBeUndefined();
-        expect(r instanceof sdk.Recognizer);
-
-        r.canceled = (o: sdk.Recognizer, e: sdk.SpeechRecognitionCanceledEventArgs) => {
-            try {
-                expect(e.errorDetails).toBeUndefined();
-                expect(sdk.CancellationErrorCode[e.errorCode]).toEqual(sdk.CancellationErrorCode[sdk.CancellationErrorCode.NoError]);
-                expect(sdk.CancellationReason[e.reason]).toEqual(sdk.CancellationReason[sdk.CancellationReason.EndOfStream]);
-            } catch (error) {
-                done.fail(error);
-            }
-        };
-        r.recognizeOnceAsync(
-            (p2: sdk.SpeechRecognitionResult) => {
-                try {
-                    const res: sdk.SpeechRecognitionResult = p2;
-
-                    expect(res).not.toBeUndefined();
-                    expect(res.reason).toEqual(sdk.ResultReason.NoMatch);
-                    const nmd: sdk.NoMatchDetails = sdk.NoMatchDetails.fromResult(res);
-                    expect(nmd.reason).toEqual(sdk.NoMatchReason.InitialSilenceTimeout);
-                    expect(res.properties).not.toBeUndefined();
-                    expect(res.properties.getProperty(sdk.PropertyId.SpeechServiceResponse_JsonResult)).not.toBeUndefined();
-
-                    done();
-                } catch (error) {
-                    done.fail(error);
-                }
-            },
-            (error: string) => {
-                done.fail(error);
-            });
-    });
-
     test("RecognizeOnceAsync is async", (done: jest.DoneCallback) => {
         // tslint:disable-next-line:no-console
         console.info("Name: ecognizeOnceAsync is async");
@@ -1709,64 +1487,6 @@ describe.each([true])("Service based tests", (forceNodeWebSocket: boolean) => {
         expect(resultSeen).toEqual(false);
         postCall = true;
     }, 100000);
-
-    test("InitialSilenceTimeout Continuous", (done: jest.DoneCallback) => {
-        // tslint:disable-next-line:no-console
-        console.info("Name: InitialSilenceTimeout Continuous");
-        const s: sdk.SpeechConfig = BuildSpeechConfig();
-        objsToClose.push(s);
-
-        let p: sdk.PullAudioInputStream;
-
-        p = sdk.AudioInputStream.createPullStream(
-            {
-                close: () => { return; },
-                read: (buffer: ArrayBuffer): number => {
-                    return buffer.byteLength;
-                },
-            });
-
-        const config: sdk.AudioConfig = sdk.AudioConfig.fromStreamInput(p);
-
-        const r: sdk.SpeechRecognizer = new sdk.SpeechRecognizer(s, config);
-        objsToClose.push(r);
-        expect(r).not.toBeUndefined();
-        expect(r instanceof sdk.Recognizer);
-
-        r.canceled = (o: sdk.Recognizer, e: sdk.SpeechRecognitionCanceledEventArgs) => {
-            // Since the pull stream above will always return an empty array, there should be
-            // no other reason besides an error for cancel to hit.
-            done.fail(e.errorDetails);
-        };
-
-        let passed: boolean = false;
-
-        r.recognized = (o: sdk.Recognizer, e: sdk.SpeechRecognitionEventArgs) => {
-
-            const res: sdk.SpeechRecognitionResult = e.result;
-            expect(res).not.toBeUndefined();
-            expect(sdk.ResultReason.NoMatch).toEqual(res.reason);
-            expect(res.text).toBeUndefined();
-
-            const nmd: sdk.NoMatchDetails = sdk.NoMatchDetails.fromResult(res);
-            expect(nmd.reason).toEqual(sdk.NoMatchReason.InitialSilenceTimeout);
-            passed = true;
-        };
-
-        /* tslint:disable:no-empty */
-        r.startContinuousRecognitionAsync(() => {
-        },
-            (error: string) => {
-                done.fail(error);
-            });
-
-        WaitForCondition(() => passed, () => {
-            r.stopContinuousRecognitionAsync(() => {
-                done();
-            }, (error: string) => done.fail(error));
-        });
-
-    }, 30000);
 
     test("Audio Config is optional", () => {
         // tslint:disable-next-line:no-console
@@ -2054,169 +1774,6 @@ describe.each([true])("Service based tests", (forceNodeWebSocket: boolean) => {
             WaitForCondition(() => (doneCount === 2), done);
         });
     });
-
-    test("Silence After Speech", (done: jest.DoneCallback) => {
-        // tslint:disable-next-line:no-console
-        console.info("Name: Silence After Speech");
-        // Pump valid speech and then silence until at least one speech end cycle hits.
-        const p: sdk.PushAudioInputStream = sdk.AudioInputStream.createPushStream();
-        const bigFileBuffer: Uint8Array = new Uint8Array(32 * 1024 * 30); // ~30 seconds.
-        const config: sdk.AudioConfig = sdk.AudioConfig.fromStreamInput(p);
-        const s: sdk.SpeechConfig = BuildSpeechConfig();
-        objsToClose.push(s);
-
-        p.write(WaveFileAudioInput.LoadArrayFromFile(Settings.WaveFile));
-        p.write(bigFileBuffer.buffer);
-        p.close();
-
-        const r: sdk.SpeechRecognizer = new sdk.SpeechRecognizer(s, config);
-        objsToClose.push(r);
-
-        let speechRecognized: boolean = false;
-        let noMatchCount: number = 0;
-        let speechEnded: number = 0;
-        let canceled: boolean = false;
-        let inTurn: boolean = false;
-
-        r.recognized = (o: sdk.Recognizer, e: sdk.SpeechRecognitionEventArgs) => {
-            try {
-                if (e.result.reason === sdk.ResultReason.RecognizedSpeech) {
-                    expect(speechRecognized).toEqual(false);
-                    speechRecognized = true;
-                    expect(sdk.ResultReason[e.result.reason]).toEqual(sdk.ResultReason[sdk.ResultReason.RecognizedSpeech]);
-                    expect(e.result.text).toEqual("What's the weather like?");
-                } else if (e.result.reason === sdk.ResultReason.NoMatch) {
-                    expect(speechRecognized).toEqual(true);
-                    noMatchCount++;
-                }
-            } catch (error) {
-                done.fail(error);
-            }
-        };
-
-        r.canceled = (o: sdk.Recognizer, e: sdk.SpeechRecognitionCanceledEventArgs): void => {
-            try {
-                expect(e.errorDetails).toBeUndefined();
-                expect(e.reason).toEqual(sdk.CancellationReason.EndOfStream);
-                canceled = true;
-            } catch (error) {
-                done.fail(error);
-            }
-        };
-
-        r.sessionStarted = ((s: sdk.Recognizer, e: sdk.SessionEventArgs): void => {
-            inTurn = true;
-        });
-
-        r.sessionStopped = ((s: sdk.Recognizer, e: sdk.SessionEventArgs): void => {
-            inTurn = false;
-        });
-
-        r.speechEndDetected = (o: sdk.Recognizer, e: sdk.RecognitionEventArgs): void => {
-            speechEnded++;
-        };
-
-        r.startContinuousRecognitionAsync(() => {
-            WaitForCondition(() => (canceled && !inTurn), () => {
-                r.stopContinuousRecognitionAsync(() => {
-                    try {
-                        expect(speechEnded).toEqual(noMatchCount);
-                        expect(noMatchCount).toEqual(2);
-                        done();
-                    } catch (error) {
-                        done.fail(error);
-                    }
-                }, (error: string) => {
-                    done.fail(error);
-                });
-            });
-        },
-            (err: string) => {
-                done.fail(err);
-            });
-    }, 30000);
-
-    test("Silence Then Speech", (done: jest.DoneCallback) => {
-        // tslint:disable-next-line:no-console
-        console.info("Name: Silence Then Speech");
-        // Pump valid speech and then silence until at least one speech end cycle hits.
-        const p: sdk.PushAudioInputStream = sdk.AudioInputStream.createPushStream();
-        const bigFileBuffer: Uint8Array = new Uint8Array(32 * 1024 * 30); // ~30 seconds.
-        const config: sdk.AudioConfig = sdk.AudioConfig.fromStreamInput(p);
-        const s: sdk.SpeechConfig = BuildSpeechConfig();
-        objsToClose.push(s);
-
-        p.write(bigFileBuffer.buffer);
-        p.write(WaveFileAudioInput.LoadArrayFromFile(Settings.WaveFile));
-        p.close();
-
-        const r: sdk.SpeechRecognizer = new sdk.SpeechRecognizer(s, config);
-        objsToClose.push(r);
-
-        let speechRecognized: boolean = false;
-        let noMatchCount: number = 0;
-        let speechEnded: number = 0;
-        let canceled: boolean = false;
-        let inTurn: boolean = false;
-
-        r.recognized = (o: sdk.Recognizer, e: sdk.SpeechRecognitionEventArgs) => {
-            try {
-                if (e.result.reason === sdk.ResultReason.RecognizedSpeech) {
-                    expect(speechRecognized).toEqual(false);
-                    expect(noMatchCount).toBeGreaterThanOrEqual(1);
-                    speechRecognized = true;
-                    expect(sdk.ResultReason[e.result.reason]).toEqual(sdk.ResultReason[sdk.ResultReason.RecognizedSpeech]);
-                    expect(e.result.text).toEqual("What's the weather like?");
-                } else if (e.result.reason === sdk.ResultReason.NoMatch) {
-                    expect(speechRecognized).toEqual(false);
-                    noMatchCount++;
-                }
-            } catch (error) {
-                done.fail(error);
-            }
-        };
-
-        r.canceled = (o: sdk.Recognizer, e: sdk.SpeechRecognitionCanceledEventArgs): void => {
-            try {
-                expect(e.errorDetails).toBeUndefined();
-                expect(e.reason).toEqual(sdk.CancellationReason.EndOfStream);
-                canceled = true;
-            } catch (error) {
-                done.fail(error);
-            }
-        };
-
-        r.sessionStarted = ((s: sdk.Recognizer, e: sdk.SessionEventArgs): void => {
-            inTurn = true;
-        });
-
-        r.sessionStopped = ((s: sdk.Recognizer, e: sdk.SessionEventArgs): void => {
-            inTurn = false;
-        });
-
-        r.speechEndDetected = (o: sdk.Recognizer, e: sdk.RecognitionEventArgs): void => {
-            speechEnded++;
-        };
-
-        r.startContinuousRecognitionAsync(() => {
-            WaitForCondition(() => (canceled && !inTurn), () => {
-                r.stopContinuousRecognitionAsync(() => {
-                    try {
-                        expect(speechEnded).toEqual(noMatchCount + 1);
-                        expect(noMatchCount).toEqual(2);
-                        done();
-                    } catch (error) {
-                        done.fail(error);
-                    }
-                }, (error: string) => {
-                    done.fail(error);
-                });
-            });
-        },
-            (err: string) => {
-                done.fail(err);
-            });
-    }, 35000);
 });
 
 Settings.testIfDOMCondition("Push Stream Async", (done: jest.DoneCallback) => {
@@ -2750,7 +2307,7 @@ describe("PhraseList tests", () => {
             (error: string) => {
                 done.fail(error);
             });
-    });
+    }, 10000);
 
     test("Phraselist Clear works.", (done: jest.DoneCallback) => {
